Make `handleNumberLiterals` public
