+++ src/builtin/main.ts
@@ -2,3 +2,4 @@ export * from './number-compare';
 export * from './number';
 export * from './string';
 export * from './wrap';
+export { handleNumberLiterals } from './util';
